+++ extension/src/content_scripts/initiate_login.js
@@ -1,0 +1,3 @@
+"use strict";
+
+window.location.replace('shib_login.php?target=');