+++ extension/src/worker.js
@@ -1,7 +1,9 @@
-chrome.runtime.onInstalled.addListener((object) => {
-	if (object.reason === chrome.runtime.OnInstalledReason.INSTALL) {
-		chrome.tabs.create({ url: "https://thechnet.github.io/adamatic" });
-	} else if (object.reason === chrome.runtime.OnInstalledReason.UPDATE) {
-		chrome.tabs.create({ url: "https://thechnet.github.io/adamatic?changes=latest" });
+"use strict";
+
+chrome.runtime.onInstalled.addListener(event => {
+	if (event.reason === chrome.runtime.OnInstalledReason.INSTALL) {
+		chrome.tabs.create({ url: 'https://thechnet.github.io/adamatic' });
+	} else if (event.reason === chrome.runtime.OnInstalledReason.UPDATE) {
+		chrome.tabs.create({ url: 'https://thechnet.github.io/adamatic?changes=2.0.0' });
 	}
 });